+++ uCOS-III_Wrapper/uCOS-III/os_flag.c
@@ -181,6 +181,8 @@ void  OSFlagCreate (OS_FLAG_GRP  *p_grp,
 *
 *                            OS_ERR_NONE                  The call was successful and the event flag group was deleted
 *                            OS_ERR_DEL_ISR               If you attempted to delete the event flag group from an ISR
+*                            OS_ERR_ILLEGAL_DEL_RUN_TIME  If you are trying to delete the event flag group after you
+*                                                            called OSStart()
 *                            OS_ERR_OBJ_PTR_NULL          If 'p_grp' is a NULL pointer.
 *                            OS_ERR_OBJ_TYPE              If you didn't pass a pointer to an event flag group
 *                            OS_ERR_OPT_INVALID           An invalid option was specified
@@ -211,6 +213,13 @@ OS_OBJ_QTY  OSFlagDel (OS_FLAG_GRP  *p_grp,
     }
 #endif
 
+#ifdef OS_SAFETY_CRITICAL_IEC61508
+    if (OSSafetyCriticalStartFlag == OS_TRUE) {
+       *p_err = OS_ERR_ILLEGAL_DEL_RUN_TIME;
+        return (0u);
+    }
+#endif
+    
 #if OS_CFG_CALLED_FROM_ISR_CHK_EN > 0u    
     if(OSIntNestingCtr > (OS_NESTING_CTR)0)/*检查是否在中断中运行*/
     {